added basic design for the layout and the simulation form. Simulation storage is working fine. Not validation yet

+++ LoanSimulation/Models/bank.cs
@@ -15,6 +15,6 @@ namespace LoanSimulation.Models
     public partial class bank
     {
         public string name { get; set; }
-        public short id { get; set; }
+        public int id { get; set; }
     }
 }
